requirements.txt
aioredis==2.0.1
aiosmtplib==4.0.1
Django==5.2.1
python-jose
fastapi==0.115.12
passlib==1.7.4
Pillow==11.2.1
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_settings==2.9.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
sse_starlette==2.3.5
starlette==0.46.2
uvicorn==0.34.2
pydantic[email]
mysql-connector-python
orjson
bcrypt==4.0.1
motor==3.3.2
pymongo==4.6.1
asyncmy
opencv-python
aiofiles
python-multipart
cryptography
alembic
celery==5.3.4
redis==5.0.1